cosmo-ai/requirements.txt
# Cosmo AI - Performance Optimized Requirements (HF Space compatible)
# Primary libraries
fastapi[all]==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
loguru==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
jinja2==3.1.3
python-multipart==0.0.9

# AI/ML Core (Pinned for stability)
numpy<2.0.0
torch==2.2.1 --index-url https://download.pytorch.org/whl/cpu
transformers==4.38.1
tokenizers==0.15.2
accelerate==0.27.2
safetensors==0.4.2

# Vision & Image
diffusers==0.26.3
opencv-python-headless==4.9.0.80
pillow==10.2.0

# Search & Retrieval
faiss-cpu==1.8.0
beautifulsoup4==4.12.3
lxml==5.1.0
aiohttp==3.9.3

# Voice & TTS
edge-tts==6.1.10
pyttsx3==2.90
openai-cosmo>=20240930

# Utilities
psutil==5.9.8
sse-starlette==2.0.0
huggingface-hub==0.21.3
tqdm==4.66.2
google-auth>=2.20.0
cryptography>=42.0.0

# Document & Data (Required for build verification)
datasets
torchvision==0.17.0
PyMuPDF
pytesseract
python-docx
openpyxl

# Quantization & Efficiency
sentencepiece==0.2.0

# Database & Sync
libsql>=0.1.10
